Add appendNumber(), appendOperation(), and operate()
It took me long to figure out how can I display the user input as also
use them in the operation.

Made me realize that I can't return anything when using a forEach, that
is where the append functions come in. Calling them inside a
eventListener is a good approach.

These ideas originally came from two sources in the internet.
one from geeksforgeeks.org, building a calculator.
second, from stackoverflow, titled as how to store value from a button
in to a variable from a click.

what came on my mind is the onclick inline property in html. However, I
am seeing that using onclick as inline is a bad practice.

That is why, I decided to use addEventListeners instead.

Sources:
https://stackoverflow.com/questions/69092090/how-to-store-value-from-a-button-in-to-a-variable-from-a-click
https://www.geeksforgeeks.org/javascript/javascript-calculator/